+++ src/iTPS/core/simple_update.cpp
@@ -35,9 +35,8 @@
 #include <mptensor/rsvd.hpp>
 #include <mptensor/tensor.hpp>
 
-#include "../../PEPS_Parameters.hpp"
-
 #include "../../tensor.hpp"
+#include "../PEPS_Parameters.hpp"
 
 namespace tenes {
