+++ src/iTPS/tensors.cpp
@@ -26,8 +26,6 @@
 
 #include "../tensor.hpp"
 
-#include "../operator.hpp"
-#include "../PEPS_Parameters.hpp"
 #include "../printlevel.hpp"
 #include "../timer.hpp"
 #include "../util/file.hpp"
